Re-organised tests
Closes #654
