+++ e2e/registration/registration.teardown.ts
@@ -1,9 +1,10 @@
 // Just import from playwright as we don't want to load the extension from the temp directory
 import { test as teardown } from '@playwright/test';
 
-import { cleanExtension } from './loader';
+import { cleanAuth, cleanExtension } from '../utils/loader';
 
-teardown('cleanup extension data', async () => {
+teardown('cleanup registration keys', async () => {
   //   Create a new page and navigate to extension
-  await cleanExtension('transaction');
+  await cleanAuth();
+  await cleanExtension('registration');
 });
