+++ e2e/transaction/transaction.setup.ts
@@ -1,5 +1,5 @@
-import { importReceiverAccount, importSenderAccount, lockExtension } from './helper';
-import { test as setup } from './loader';
+import { importReceiverAccount, importSenderAccount } from '../utils/helper';
+import { test as setup } from '../utils/loader';
 
 setup('Import sender and receiver accounts', async ({ page, extensionId }) => {
   // Lock the extension and import sender and receiver accounts
