+++ e2e/transaction/transaction.teardown.ts
@@ -1,10 +1,9 @@
 // Just import from playwright as we don't want to load the extension from the temp directory
 import { test as teardown } from '@playwright/test';
 
-import { cleanAuth, cleanExtension } from './loader';
+import { cleanExtension } from '../utils/loader';
 
-teardown('cleanup registration keys', async () => {
+teardown('cleanup extension data', async () => {
   //   Create a new page and navigate to extension
-  await cleanAuth();
-  await cleanExtension('registration');
+  await cleanExtension('transaction');
 });
