--- e2e/utils/registration.setup.ts
@@ -1,6 +1,0 @@
-import { test as setup } from './loader';
-
-// for user register and login
-setup('setup new wallet or login if already registered', async ({ page, extensionId }) => {
-  await page.goto(`chrome-extension://${extensionId}/index.html#/dashboard`);
-});
